Split Modal parts
