+++ packages/fuselage/src/components/Modal/ModalBackdrop.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+import { Box } from '../Box';
+
+export const ModalBackdrop = (props) => <Box rcx-modal__backdrop {...props} />;
