+++ packages/fuselage/src/components/Modal/ModalClose.js
@@ -1,0 +1,10 @@
+import React from 'react';
+
+import { Button } from '../Button';
+import { Icon } from '../Icon';
+
+export const ModalClose = (props) => (
+  <Button small ghost flexShrink={0} {...props}>
+    <Icon name='cross' size='x24' />
+  </Button>
+);
