+++ packages/fuselage/src/components/Modal/ModalFooter.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+import { Box } from '../Box';
+
+export const ModalFooter = ({ children }) => (
+  <Box rcx-modal__footer margin={32}>
+    {children}
+  </Box>
+);
