+++ packages/fuselage/src/components/Modal/ModalThumb.js
@@ -1,0 +1,10 @@
+import React from 'react';
+
+import { Avatar } from '../Avatar';
+import { Box } from '../Box';
+
+export const ModalThumb = (props) => (
+  <Box>
+    <Avatar size='x32' {...props} />
+  </Box>
+);
